Keep FiltersModal in separate component
