+++ front/app/components/IdeaCards/IdeasWithFiltersSidebar/ButtonWithFiltersModal/FiltersModal/BottomBar.tsx
@@ -8,7 +8,7 @@ import Button from 'components/UI/Button';
 
 import { FormattedMessage } from 'utils/cl-intl';
 
-import { Props as InputFiltersProps } from '../InputFilters';
+import { Props as InputFiltersProps } from '../../InputFilters';
 
 import messages from './messages';
