+++ front/app/components/IdeaCards/IdeasWithFiltersSidebar/ButtonWithFiltersModal/FiltersModal/TopBar.tsx
