+++ front/app/components/IdeaCards/IdeasWithFiltersSidebar/ButtonWithFiltersModal/index.tsx
@@ -1,20 +1,13 @@
 import React, { useState } from 'react';
 
-import { Box, Button, useBreakpoint } from '@citizenlab/cl2-component-library';
-import { isNumber } from 'lodash-es';
-
-import useIdeasFilterCounts from 'api/ideas_filter_counts/useIdeasFilterCounts';
-
-import FullscreenModal from 'components/UI/FullscreenModal';
+import { Button, useBreakpoint } from '@citizenlab/cl2-component-library';
 
 import { FormattedMessage } from 'utils/cl-intl';
-import { isNilOrError } from 'utils/helperUtils';
 
-import InputFilters, { Props as InputFiltersProps } from '../InputFilters';
+import { Props as InputFiltersProps } from '../InputFilters';
 
-import BottomBar from './BottomBar';
-import messages from './messages';
-import TopBar from './TopBar';
+import FiltersModal from './FiltersModal';
+import messages from './FiltersModal/messages';
 
 const ButtonWithFiltersModal = ({
   selectedIdeaFilters,
@@ -23,11 +16,6 @@ const ButtonWithFiltersModal = ({
 }: InputFiltersProps) => {
   const isSmallerThanTablet = useBreakpoint('tablet');
   const [filtersModalOpened, setFiltersModalOpened] = useState(false);
-
-  const { data: ideasFilterCounts } = useIdeasFilterCounts(selectedIdeaFilters);
-  const total = isNilOrError(ideasFilterCounts)
-    ? null
-    : ideasFilterCounts.data.attributes.total;
 
   const openModal = () => {
     setFiltersModalOpened(true);
@@ -49,39 +37,13 @@ const ButtonWithFiltersModal = ({
         mb="12px"
         mt="4px"
       />
-      <FullscreenModal
+      <FiltersModal
         opened={filtersModalOpened}
-        close={closeModal}
-        topBar={<TopBar onReset={onClearFilters} onClose={closeModal} />}
-        bottomBar={
-          <BottomBar
-            buttonText={
-              total && isNumber(total) ? (
-                <FormattedMessage
-                  {...messages.showXResults}
-                  values={{
-                    ideasCount: total,
-                  }}
-                />
-              ) : (
-                <FormattedMessage {...messages.showResults} />
-              )
-            }
-            onClick={closeModal}
-          />
-        }
-        contentBgColor="background"
-      >
-        <Box p="16px">
-          <InputFilters
-            selectedIdeaFilters={selectedIdeaFilters}
-            onClearFilters={onClearFilters}
-            // We have a reset filters button in TopBar
-            showResetButton={false}
-            {...filtersProps}
-          />
-        </Box>
-      </FullscreenModal>
+        selectedIdeaFilters={selectedIdeaFilters}
+        onClearFilters={onClearFilters}
+        onClose={closeModal}
+        {...filtersProps}
+      />
     </>
   );
 };
